pyproject.toml
[project]
name = "train"
version = "0.1.0"
description = "Car coupling detector"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "pi-heif>=1.3.0",
  "pillow>=12.2.0",
  "torch>=2.0.0",
  "torchvision>=0.15.0",
  "ultralytics>=8.0.0",
]

[project.scripts]
train = "train.train:main"
predict = "train.predict:main"

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"
